Saving products bugfix

+++ src/TD.Web/TD.Web.Admin/TD.Web.Admin.Contracts/Requests/Products/ProductsSaveRequest.cs
@@ -9,14 +9,14 @@ public class ProductsSaveRequest : LSCoreSaveRequest
     public string? Src { get; set; }
     public string Image { get; set; }
     public string? CatalogId { get; set; }
-    public int UnitId { get; set; }
+    public long UnitId { get; set; }
     public ProductClassification Classification { get; set; }
     public decimal VAT { get; set; }
     public List<long> Groups { get; set; } = [];
-    public int ProductPriceGroupId { get; set; }
+    public long ProductPriceGroupId { get; set; }
     public string? Description { get; set; }
     public string? ShortDescription { get; set; }
-    public int? AlternateUnitId { get; set; }
+    public long? AlternateUnitId { get; set; }
     public decimal? OneAlternatePackageEquals { get; set; }
     public int PriorityIndex { get; set; }
     public string? MetaTitle { get; set; }
